requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-WTF==1.1.1
frozenlist==1.3.3
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
python-dotenv==1.0.0
python-graphql-client==0.4.3
requests==2.30.0
urllib3==2.0.2
websockets==11.0.3
Werkzeug==2.3.4
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
